% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadm}
\alias{gadm}
\alias{gadm0}
\alias{gadm0r}
\alias{gadm1}
\alias{gadm1r}
\title{GADM maps for Vietnam}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 68 columns.}
\source{
GADM data base from \url{www.gadm.org}.
}
\usage{
gadm0

gadm0r

gadm1

gadm1r
}
\description{
The GADM maps of the country and provinces borders of Vietnam.
}
\details{
\code{gadm0} and \code{gadm0ir} are the GADM maps of the country border and
\code{gadm1} and \code{gadm1r} are the GADM maps of the provinces borders.
\code{gadm0r} and \code{gadm1r} are the raw GADM objects (heavy in memory)
whereas \code{gadm0} and \code{gadm1} are the "thinned" GADM objects for
easier plotting.
}
\examples{
## Plotting the country border of Vietnam:
plot(gadm0)
## Adding the provinces borders:
plot(gadm1,add=T)
}
\author{
Marc Choisy
}
\keyword{datasets}

