% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadmVN.R
\docType{package}
\name{gadmVN}
\alias{gadmVN}
\alias{gadmVN-package}
\title{gadmVN: A database gathering vector and raster maps of Vietnam.}
\description{
The mapsVN package provides two maps of Vietnam: gadm0 and gadm1.
}

