% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm.R
\name{gadm}
\alias{gadm}
\title{Maps of Vietnam}
\source{
GADM data base from \url{www.gadm.org}.
}
\usage{
gadm(date = "2015-01-01", level = "provinces", resolution = "low")
}
\arguments{
\item{date}{Text in the "YYYY-MM-DD" format. For example "2015-01-17" for the
17th of January 2017. By default "2015-01-01".}

\item{level}{Text ("country" or "provinces"). By default "provinces".}

\item{resolution}{Text ("low" or "high"). By default "low".}
}
\description{
Generates a country or provinces map of Vietnam for a given date and
resolution.
}
\details{
This function generates a country or provinces map of Vietnam at a given date.
Indeed the delimitations of Vietnamese provinces have changed through the
history with the number of provinces increasing from 40 in 1979 to 63 in 2008.
Most of the events are province splits and one merging.
}
\author{
Marc Choisy
}

