% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadmVN.R
\docType{package}
\name{gadmVN}
\alias{gadmVN}
\alias{gadmVN-package}
\title{gadmVN: Country and provinces maps of Vietnam}
\description{
The mapsVN package provides country and provinces maps of Vietnam at
different points in time.
}

